create_schema.py - factor out schema creation into a separate file
